requirements.txt
altgraph==0.17.4
anyio==4.10.0
async-timeout==5.0.1
betterproto==2.0.0b7
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
ffmpy==0.6.1
Flask==3.1.1
grpclib==0.4.8
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mashumaro==3.16
multidict==6.6.4
packaging==25.0
pefile==2023.2.7
protobuf==6.32.0
protobuf3-to-dict==0.1.5
pyee==13.0.0
pygame==2.6.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-socks==2.7.2
pywin32==311
pywin32-ctypes==0.2.3
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
TikTokLive==6.6.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
websockets_proxy==0.1.3
Werkzeug==3.1.3
